src/api/cs3/requirements.txt
numpy
opencv-python
tensorflow==1.8.0
Pillow
pathlib
keras
matplotlib
sklearn
easydict